+++ lib/ui/page/home/page/chat/widget/chat_forward.dart
@@ -1136,19 +1136,15 @@ class _ChatForwardWidgetState extends State<ChatForwardWidget> {
                                   variants: [
                                     if (!deletable || !isMonolog)
                                       ConfirmDialogVariant(
+                                        key: const Key('HideForMe'),
                                         onProceed: widget.onHide,
-                                        child: Text(
-                                          'label_delete_for_me'.l10n,
-                                          key: const Key('HideForMe'),
-                                        ),
+                                        label: 'label_delete_for_me'.l10n,
                                       ),
                                     if (deletable)
                                       ConfirmDialogVariant(
+                                        key: const Key('DeleteForAll'),
                                         onProceed: widget.onDelete,
-                                        child: Text(
-                                          'label_delete_for_everyone'.l10n,
-                                          key: const Key('DeleteForAll'),
-                                        ),
+                                        label: 'label_delete_for_everyone'.l10n,
                                       )
                                   ],
                                 );