+++ lib/ui/page/home/page/my_profile/widget/login.dart
@@ -123,6 +123,7 @@ class _UserLoginFieldState extends State<UserLoginField> {
             ),
       label: 'label_login'.l10n,
       hint: widget.login == null ? 'label_login_hint'.l10n : widget.login!.val,
+      clearable: false,
       subtitle: RichText(
         text: TextSpan(
           children: [
@@ -159,15 +160,15 @@ class _UserLoginFieldState extends State<UserLoginField> {
                     variants: [
                       ConfirmDialogVariant(
                         onProceed: () {},
-                        child: Text('label_all'.l10n),
+                        label: 'label_all'.l10n,
                       ),
                       ConfirmDialogVariant(
                         onProceed: () {},
-                        child: Text('label_my_contacts'.l10n),
+                        label: 'label_my_contacts'.l10n,
                       ),
                       ConfirmDialogVariant(
                         onProceed: () {},
-                        child: Text('label_nobody'.l10n),
+                        label: 'label_nobody'.l10n,
                       ),
                     ],
                   );