+++ CHANGELOG.md
@@ -19,13 +19,23 @@ All user visible changes to this project will be documented in this file. This p
         - Redesigned typography tab. ([#663], [#615])
     - Chat page:
         - Display message field while loading. ([#662], [#634])
+    - Disabled larger fonts accessibility setting temporary. ([#679])
 - Web:
     - Updated loading animation. ([#662], [#634])
 
+### Fixed
+
+- UI:
+    - Profile page:
+        - Save button displaying when login field is empty. ([#672], [#575])
+
 [#615]: /../../issues/615
 [#634]: /../../issues/634
+[#575]: /../../issues/575
 [#662]: /../../pull/662
 [#663]: /../../pull/663
+[#672]: /../../pull/672
+[#679]: /../../pull/679